WL-3038 Permissions helper: contribute role sees only a blank screen.
This is a combination of 2 commits.

WL-3038 Display nice message when you can't edit permissions.

We make sure that when you don't have permission to edit the permissions you get a nice message. We also make sure that when something else goes wrong you end up at a proper page rather than a blank screen and have some way of getting back.

WL-3038 Update other languages for property rename.

We dropped the java. prefix as it's no longer used in the Java code.