WL-3191 Also include roles for content
 This is a combination of 4 commits.

WL-3191 Detect type of ROLES_REF

We are now allowing collections of roles_refs to be defined but we should
still support a String.

Internally these will all be modelled as a Collection now.

WL-3191 Pluralise some of the properties

WL-3191 Just take the first role for wierd check

I'm not sure why this part of the code is executed or what it even does.

Matthew and I agree it shouldn't really be there and will probably result
in buggy behaviour.

The code is probably hit rarely enough that it doesn't matter, so instead
of converting this code to handle the new Collection type, just take the
first element and use that.

WL-3191 Add all the roles uniquely

Now that we are adding roles from multiple realms, let's make sure they
only appear once each.