WL-3191 Just take the first role for wierd check
I'm not sure why this part of the code is executed or what it even does.

Matthew and I agree it shouldn't really be there and will probably result
in buggy behaviour.

The code is probably hit rarely enough that it doesn't matter, so instead
of converting this code to handle the new Collection type, just take the
first element and use that.